script: move element interfaces into `element/`
It moves `Element.rs` into a subdirectory and turns on
dead code detection (found 4 methods that could be removed).
Then it also extracts the first attribute-related method
into a separate file, to reduce the file size of `Element`
and add more structure as part of #43709.

Part of #38901
Part of #43709
